rename ChooseOptions component to VotePreference
