--- app/components/views/ProposalDetailsPage/helpers/ChooseOptions/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ChooseOptions";
