+++ app/components/views/ProposalDetailsPage/helpers/VotePreference/VotedCheckmark/VotedCheckmark.module.css
