+++ app/components/views/ProposalDetailsPage/helpers/VotePreference/VotedCheckmark/index.js
