+++ app/components/views/ProposalDetailsPage/helpers/VotePreference/index.js
@@ -1,0 +1,1 @@
+export { default } from "./VotePreference";
